Added count nodes feature
It may show that a lot of changes has been made because i took your code and edited it in my IDE and then pasted it again, but the only change that i made was adding the 'count nodes' feature. I've checked it and it is working fine. Please merge.Thanks